fix eslint and delete useStickyState file to implement in the future
